+++ keeper/keeper.go
@@ -8,7 +8,7 @@ type ReqTask struct {
 }
 
 type RespTask struct {
-	ReqId   int      `json:"id"`
+	ReqId   uint64   `json:"id"`
 	Status  int      `json:"status"`
 	Headers []Header `json:"headers"`
 	Length  int      `json:"length"`
@@ -19,15 +19,16 @@ type Header struct {
 	Value string `json:"value"`
 }
 
+// структура для записи значения в мапу
 type Task struct {
 	ReqTask  ReqTask
 	RespTask RespTask
 }
 
 type Keeper interface {
-	SaveTask(task Task) int
-	UpdateTask(reqId int, task Task)
-	GetById(id int) (task Task, ok bool)
+	SaveTask(task Task) uint64
+	UpdateTask(reqId uint64, task Task)
+	GetById(id uint64) (task Task, ok bool)
 	GetAll() []ReqTask
-	DeleteById(id int)
+	DeleteById(id uint64)
 }
